[OVERKIZ] led: add Cycle LED trigger
Currently, none of available triggers supports playing with the LED brightness
level.  The cycle trigger provides a way to define custom brightness cycle.
For example, it is easy to customize the cycle to mock up the rhythm of human
breathing which is a nice cycle to tell the user the system is doing something.

This trigger is meant to be usable for waiting an event to happen, for example
when the system gets ready.  Those cycles may be used to reflect well known
system status (e.g. idle mode, startup...).

This implementation provides several interfaces:
 - to define the cycle itself:
   * plot: definition of plot points using plot or rawplot files,
           each points defines the brightness level
   * interval: constant time interval between each plot point
 - to control the cycle:
   * repeat: the number of repetition of the whole plot cycle
             0 for an infinite loop
   * control: used to control the cycle trigger
     + "start"/"stop": to start/stop the cycle
     + "reset" to clear the cycle counter and the internal plot point index
     + "pause"/"resume" to pause/resume the cycle

Change-Id: I3ccae77a45453504253c89f673c1c9a94fce230c
Signed-off-by: Gaël PORTAY <gael.portay@gmail.com>